requirements.txt
-f https://download.pytorch.org/whl/cu118
cog~=0.8.6
dlib~=19.24.1
lmdb~=1.4.0
lpips~=0.1.4
matplotlib~=3.7.1
monai[nibabel, itk]~=1.2.0
numpy~=1.24.4
pandas~=1.1.5
Pillow~=10.0.0
pytorch_fid~=0.2.1
pytorch_lightning~=2.0.6
PyYAML~=6.0.1
requests~=2.31.0
scipy~=1.11.1
seaborn~=0.12.2
setuptools~=68.0.0
tensorboard~=2.14.0
timm~=0.9.2
torch~=2.0.1
torchmetrics~=1.0.3
torchvision~=0.15.2
tqdm~=4.66.0
wandb~=0.15.0
